pyproject.toml
[project]
name = "python-drf-backend-exam"
version = "0.1.0"
description = ""
authors = [
    {name = "Donato4Humans",email = "donato.4work@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "djangorestframework (>=3.16.0,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "django-filter (>=25.1,<26.0)",
    "django-celery-beat (>=2.8.1,<3.0.0)",
    "django-celery-results (>=2.6.0,<3.0.0)",
    "daphne (>=4.2.1,<5.0.0)",
    "channels-redis (>=4.2.1,<5.0.0)",
    "drf-yasg (>=1.21.10,<2.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "djangochannelsrestframework (>=1.3.0,<2.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "pymysql (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
isort = "^6.0.1"

